# DateTime: Use Optimised Equals Implementation in object.Equals()