--- a/src/libraries/System.Private.CoreLib/src/System/DateTime.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/DateTime.cs
@@ -1170,11 +1170,7 @@ namespace System
         //
         public override bool Equals([NotNullWhen(true)] object? value)
         {
-            if (value is DateTime time)
-            {
-                return this == time;
-            }
-            return false;
+            return value is DateTime dt && this == dt;
         }
 
         public bool Equals(DateTime value)